timeline/serializers.py
from rest_framework import serializers
from accounts.models import User
from .models import Post

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = ['title','context']