requirements.txt
Django~=4.1.3
django-allauth~=0.41.0
django-widget-tweaks~=1.4.12
djangorestframework~=3.14.0

